Rename 'services' folder to 'common' to avoid confusion with 'streaming-services'
