+++ src/common/BrowserAction.ts
