+++ src/common/BrowserStorage.ts
