+++ src/common/Errors.ts
