+++ src/common/Session.ts
