+++ src/common/Shared.ts
