+++ src/global.d.ts
@@ -1,6 +1,6 @@
 declare interface Window {
 	wrappedJSObject?: {
-		fetch: import('./services/Requests').Fetch;
+		fetch: import('./common/Requests').Fetch;
 		fetchOptions: Record<string, unknown>;
 		netflix?: import('./streaming-services/netflix/NetflixApi').NetflixGlobalObject;
 		sdk?: import('./streaming-services/hbo-go/HboGoApi').HboGoGlobalObject;
