+++ src/modules/content/trakt/trakt.ts
@@ -1,4 +1,4 @@
-import { Session } from '../../../services/Session';
+import { Session } from '../../../common/Session';
 
 const init = async () => {
 	await Session.finishLogin();
