+++ src/modules/history/HistoryApp.tsx
@@ -6,8 +6,8 @@ import { Redirect, Route, Router, Switch } from 'react-router-dom';
 import { ErrorBoundary } from '../../components/ErrorBoundary';
 import { UtsDialog } from '../../components/UtsDialog';
 import { UtsSnackbar } from '../../components/UtsSnackbar';
-import { EventDispatcher } from '../../services/Events';
-import { Session } from '../../services/Session';
+import { EventDispatcher } from '../../common/Events';
+import { Session } from '../../common/Session';
 import { HistoryHeader } from './components/HistoryHeader';
 import { AboutPage } from './pages/AboutPage';
 import { HomePage } from './pages/HomePage';
