+++ src/modules/history/components/history/HistoryListItem.tsx
@@ -3,7 +3,7 @@ import { green, red } from '@material-ui/core/colors';
 import SyncIcon from '@material-ui/icons/Sync';
 import * as React from 'react';
 import { Item } from '../../../../models/Item';
-import { EventDispatcher } from '../../../../services/Events';
+import { EventDispatcher } from '../../../../common/Events';
 import { StreamingServiceId } from '../../../../streaming-services/streaming-services';
 import { HistoryListItemCard } from './HistoryListItemCard';
