+++ src/modules/history/components/history/HistoryOptionsList.tsx
@@ -1,6 +1,6 @@
 import { Button, ButtonGroup, FormGroup, Typography } from '@material-ui/core';
 import * as React from 'react';
-import { SyncOption } from '../../../../services/BrowserStorage';
+import { SyncOption } from '../../../../common/BrowserStorage';
 import { SyncStore } from '../../../../streaming-services/common/SyncStore';
 import { HistoryOptionsListItem } from './HistoryOptionsListItem';
