+++ src/modules/history/components/history/HistoryOptionsListItem.tsx
@@ -1,7 +1,7 @@
 import { FormControlLabel, Switch, TextField } from '@material-ui/core';
 import * as React from 'react';
-import { SyncOption } from '../../../../services/BrowserStorage';
-import { EventDispatcher } from '../../../../services/Events';
+import { SyncOption } from '../../../../common/BrowserStorage';
+import { EventDispatcher } from '../../../../common/Events';
 
 interface HistoryOptionsListItemProps {
 	option: SyncOption;
