+++ src/modules/options/options.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import '../../assets/assets';
-import { BrowserStorage } from '../../services/BrowserStorage';
-import { Errors } from '../../services/Errors';
-import { Shared } from '../../services/Shared';
+import { BrowserStorage } from '../../common/BrowserStorage';
+import { Errors } from '../../common/Errors';
+import { Shared } from '../../common/Shared';
 import { OptionsApp } from './OptionsApp';
 
 const init = async () => {
