+++ src/modules/popup/PopupApp.tsx
@@ -4,8 +4,8 @@ import * as React from 'react';
 import { useEffect, useState } from 'react';
 import { Redirect, Route, Router, Switch } from 'react-router-dom';
 import { ErrorBoundary } from '../../components/ErrorBoundary';
-import { EventDispatcher } from '../../services/Events';
-import { Session } from '../../services/Session';
+import { EventDispatcher } from '../../common/Events';
+import { Session } from '../../common/Session';
 import { PopupHeader } from './components/PopupHeader';
 import { AboutPage } from './pages/AboutPage';
 import { HomePage } from './pages/HomePage';
