+++ src/modules/popup/components/PopupHeader.tsx
@@ -3,8 +3,8 @@ import { History } from 'history';
 import * as PropTypes from 'prop-types';
 import * as React from 'react';
 import { UtsLeftRight } from '../../../components/UtsLeftRight';
-import { Session } from '../../../services/Session';
-import { Tabs } from '../../../services/Tabs';
+import { Session } from '../../../common/Session';
+import { Tabs } from '../../../common/Tabs';
 
 interface IPopupHeader {
 	history: History;
