+++ src/modules/popup/pages/AboutPage.tsx
@@ -1,6 +1,6 @@
 import { Button, Typography } from '@material-ui/core';
 import * as React from 'react';
-import { Tabs } from '../../../services/Tabs';
+import { Tabs } from '../../../common/Tabs';
 import { PopupInfo } from '../components/PopupInfo';
 
 export const AboutPage: React.FC = () => {
