+++ src/modules/popup/pages/HomePage.tsx
@@ -3,8 +3,8 @@ import * as React from 'react';
 import { useEffect, useState } from 'react';
 import { useHistory } from 'react-router-dom';
 import { UtsCenter } from '../../../components/UtsCenter';
-import { BrowserStorage } from '../../../services/BrowserStorage';
-import { Session } from '../../../services/Session';
+import { BrowserStorage } from '../../../common/BrowserStorage';
+import { Session } from '../../../common/Session';
 import { PopupNotWatching } from '../components/PopupNotWatching';
 import { PopupWatching } from '../components/PopupWatching';
 import { TraktItem } from '../../../models/TraktItem';
